jd_ai_beautify/jd-beautifier-ai-fastapi/fastapi_streamlit_app/sqlalchemy/app/requirements.txt
fastapi==0.110.1
starlette==0.37.2
uvicorn==0.29.0
asyncpg==0.29.0
beanie==1.25.0
motor==3.4.0
pymongo==4.7.0
sqlalchemy==2.0.29
fastapi-users==13.0.0
fastapi-users-db-beanie==3.0.0
fastapi-users-db-sqlalchemy==6.0.1
aiosqlite==0.20.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
bcrypt==4.1.2
passlib==1.7.4
pyjwt==2.8.0
cryptography==42.0.5
email-validator==2.1.1
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.6.4
dataclasses-json==0.6.5
httpx==0.27.0
httpx-oauth==0.14.1
async-timeout==4.0.3
tenacity==8.2.3
openai==1.26.0